refactor manager tests
